server/handler/handler.go
package handler

import (
	"errors"
	"github.com/aws/aws-lambda-go/events"
	"log"
)

var ErrNameNotProvided = errors.New("no name was provided in the HTTP body")

func Handler(request events.APIGatewayProxyRequest) (events.APIGatewayProxyResponse, error) {
	log.Printf("Processing Lambda request %s\n", request.RequestContext.RequestID)

	if len(request.Body) < 1 {
		return events.APIGatewayProxyResponse{}, ErrNameNotProvided
	}

	return events.APIGatewayProxyResponse{
		Body:       "Hello " + request.Body,
		StatusCode: 200,
	}, nil
}
